Refactor notification service and add critical alerts
The notificationService.js file has been refactored from notificationsService.js. Additionally, a new functionality has been added to monitor the number of orders waiting for payment. If this number exceeds a certain threshold (as defined in constants.js file), a critical notification is dispatched. This change will help in closely monitoring and taking potential actions for orders that are waiting for payments.